style : 명명 규칙 반영
style : 명명 규칙 반영

- user.tsx => model.user.tsx 로 파일명 변경
- ESlint 적용
